Further improved errorhandling

+++ Controllers/SuppliersController.cs
@@ -1,14 +1,11 @@
 ﻿using AutoMapper;
-using Microsoft.Ajax.Utilities;
 using NorthwindAPI.Data;
 using NorthwindAPI.Data.Entities;
 using NorthwindAPI.Models;
-using System;
 using System.Collections.Generic;
 using System.Net;
-using System.Net.Http;
 using System.Threading.Tasks;
-using System.Web;
+using System.Web.Configuration;
 using System.Web.Http;
 
 namespace NorthwindAPI.Controllers
@@ -26,25 +23,40 @@ namespace NorthwindAPI.Controllers
 
         public async Task<IHttpActionResult> Get()
         {
+            try
+            {
                 var result = await _repository.GetAllSuppliersAsync();
                 if (result != null)
                 {
                     var mappedResult = _mapper.Map<IEnumerable<SupplierModel>>(result);
                     return Ok(mappedResult);
                 }
-                else return NotFound();
+                else return Content(HttpStatusCode.NotFound, "No suppliers could be found");;
+            }
+            catch
+            {
+                return InternalServerError();
+            }
         }
 
         [ActionName("getSingleSupplierById")]
         public async Task<IHttpActionResult> GetSingleAsync(int id)
         {
+            try
+            {
                 var result = await _repository.GetSuppliersAsync(id);
                 if (result != null)
                 {
                     var mappedResult = _mapper.Map<SupplierModel>(result);
                     return Ok(mappedResult);
                 }
-                else return NotFound();
+                else return Content(HttpStatusCode.NotFound, string.Format("Supplier with id {0} could not be found", id));
+            }
+            catch
+            {
+                return InternalServerError();
+            }
+
         }
 
         [ActionName("deleteSupplierById")]
@@ -53,10 +65,14 @@ namespace NorthwindAPI.Controllers
             try
             {
                 var supplier = await _repository.GetSuppliersAsync(id);
-                _repository.DeleteSupplier(supplier);
-                if (await _repository.SaveChangesAsync())
-                    return Ok(string.Format("Supplier with id {0} has been deleted", id));
-                else return BadRequest();
+                if (supplier != null)
+                {
+                    _repository.DeleteSupplier(supplier);
+                    if (await _repository.SaveChangesAsync())
+                        return Ok(string.Format("Supplier with id {0} has been deleted", id));
+                    else return Content(HttpStatusCode.Conflict, string.Format("Supplier with id {0} could not be deleted", id));
+                }
+                else return Content(HttpStatusCode.NotFound, string.Format("Supplier with id {0} could not be found", id));
             }
             catch
             {
@@ -78,13 +94,14 @@ namespace NorthwindAPI.Controllers
                         var newModel = _mapper.Map<SupplierModel>(supplier);
                         return CreatedAtRoute("", "", newModel);
                     }
+                    else return Content(HttpStatusCode.Conflict, string.Format("Supplier with id {0} could not be updated", supplier.SupplierID));
                 }
+                else return Content(HttpStatusCode.BadRequest, "Incorrect input data");
             }
             catch
             {
                 return InternalServerError();
             }
-            return BadRequest();
         }
         
         [ActionName("putSupplierById")]
@@ -97,16 +114,20 @@ namespace NorthwindAPI.Controllers
                     model.SupplierID = id;
                     var supplier = await _repository.GetSuppliersAsync(id);
 
-                    _mapper.Map(model, supplier);
-                    _repository.UpdateSupplier(supplier);
-                    if (await _repository.SaveChangesAsync())
+                    if (supplier != null)
                     {
-                        var newModel = _mapper.Map<SupplierModel>(supplier);
-                        return Ok(newModel);
+                        _mapper.Map(model, supplier);
+                        _repository.UpdateSupplier(supplier);
+                        if (await _repository.SaveChangesAsync())
+                        {
+                            var newModel = _mapper.Map<SupplierModel>(supplier);
+                            return Ok(newModel);
+                        }
+                        else return Content(HttpStatusCode.Conflict, string.Format("Supplier with id {0} could not be updated", supplier.SupplierID));
                     }
-                    else return NotFound();
+                    else return Content(HttpStatusCode.Conflict, string.Format("Supplier with id {0} could not be found", supplier.SupplierID));
                 }
-                else return BadRequest();
+                else return Content(HttpStatusCode.BadRequest, "Incorrect input data");
             }
             catch
             {
